Lib and import changes (#21)
* deleted all bin files due to name change

* bin_tests -> lib_tests

* updated with new package name

* updated with new package name in imports

* changed package name to lib

* changed files moved to new folder

* bin files moved to lib

* updated pre-written .gitignore to not ignore lib/

* response_handler import changed

* authHanler imported directly instead of a package import.

* pep changes
