--- bin/handlers/__init__.py
@@ -1,5 +1,0 @@
-from .github_api_handler import *
-from .response_handler import *
-from .exception_handler import *
-from .authHandler import *
-from .files_handler import *
