+++ lib/__init__.py
@@ -1,5 +1,9 @@
 # __init__.py
-
+"""
+This is package contains all API and Exception specific code.
+This library has functions that interact with external API's
+Thus library also has functions and modules that interact with Github API
+"""
 from .utils import *
 from .output import *
