+++ lib/handlers/__init__.py
@@ -1,0 +1,9 @@
+"""
+Contains modules for handling different and specific API's.
+Also has Classes to hold API responses and payloads and app specific exception handling.
+"""
+from .github_api_handler import *
+from .response_handler import *
+from .exception_handler import *
+from .authHandler import *
+from .files_handler import *
