+++ lib/handlers/authHandler.py
@@ -13,7 +13,7 @@ def get_hash_key(key: str, req: func.HttpRequest) -> str:
 
 def generate_jwt_token(payload: dict) -> str:
     file_path: str = 'D:/work_src/Medium Coders Asylum/bots/cert2.pem'
-    pwd: str = None
+    pwd: bytes = None
     with open(file_path, 'rb') as file:
         _pvt_key = serialization.load_pem_private_key(data=file.read(), password=pwd, backend=default_backend)
     return encode(payload=payload, key=_pvt_key, algorithm='RS256')
