+++ lib/output.py
@@ -1,4 +1,3 @@
-import subprocess
 from time import sleep
