+++ secretaryBot/__init__.py
@@ -1,6 +1,10 @@
+# __init__.py
+"""
+This is the Aatmaram Bhide Secretary Bot.
+"""
 from .app import *
 
-# ./src files
-from .src.jobs import *
-from .src.webhook import *
-from .src.initalise import *
+# # ./src files
+# from .src.jobs import *
+# from .src.webhook import *
+# from .src.initalise import *
