+++ secretaryBot/src/__init__.py
@@ -1,5 +1,7 @@
-# __init_.py
-
+# __init__.py
+"""
+Source files for Secretary Bot.
+"""
 from .jobs import *
 from .webhook import *
 from .initalise import *
