+++ secretaryBot/src/jobs.py
@@ -2,8 +2,10 @@ from azure import functions as func
 
 
 class jobs:
-    def __init__(self, req: func.HttpRequest) -> None:
-        pass
+    _req: func.HttpRequest
+
+    def __init__(self, req: func.HttpRequest):
+        self._req = req
 
     def map_request(self):
         pass
