+++ tests/lib_tests/handlers_test/test_response_handler.py
