+++ tests/lib_tests/output_test.py
@@ -2,7 +2,7 @@ from unittest import TestCase, main
 from io import StringIO
 from unittest.mock import patch
 
-from bin import Output
+from lib import Output
 
 
 class unittests(TestCase):
